+++ kgdata/wikidata/models/wdproperty.py
@@ -3,8 +3,10 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Literal, Mapping
 
+from kgdata.models.multilingual import MultiLingualString, MultiLingualStringList
 from kgdata.models.ont_property import OntologyProperty
 from kgdata.wikidata.models.wdentity import WDEntity
+from kgdata.wikidata.models.wdstatement import WDStatement
 
 # wikibase-lexeme, monolingualtext, wikibase-sense, url, wikibase-property,
 # wikibase-form, external-id, time, commonsMedia, quantity, wikibase-item, musical-notation,
@@ -30,9 +32,10 @@ WDDataType = Literal[
 ]
 
 
-@dataclass
+@dataclass(kw_only=True, slots=True)
 class WDProperty(OntologyProperty):
     datatype: WDDataType
+    constraints: list[WDStatement]
 
     @staticmethod
     def from_entity(ent: WDEntity):
@@ -56,10 +59,49 @@ class WDProperty(OntologyProperty):
             else:
                 assert False, f"Unknown type: {stmt.value.to_dict()}"
 
-        subjects = []
-        for stmt in ent.props.get("P1629", []):
-            assert stmt.value.is_entity_id(stmt.value)
-            subjects.append(stmt.value.as_entity_id())
+        constraints = ent.props.get("P2302", [])
+        domains = None
+        ranges = None
+        for stmt in constraints:
+            entid = stmt.value.as_entity_id_safe()
+            # subject type constraint
+            if entid == "Q21503250":
+                try:
+                    # domains so it must have class -- if not, it's bad and we can ignore
+                    if "P2308" not in stmt.qualifiers:
+                        continue
+                    # and the relation must be instanceof or (instanceof or subclassof), or subclassof
+                    assert "P2309" in stmt.qualifiers, (ent.id, stmt)
+                    relations = [
+                        x.as_entity_id_safe() for x in stmt.qualifiers["P2309"]
+                    ]
+                    for relation in relations:
+                        assert relation in ["Q21503252", "Q30208840", "Q21514624"], (
+                            ent.id,
+                            stmt,
+                        )
+                except:
+                    continue
+                domains = [x.as_entity_id_safe() for x in stmt.qualifiers["P2308"]]
+
+            # value-type constraint
+            if entid == "Q21510865":
+                try:
+                    # if ranges are classes
+                    assert "P2308" in stmt.qualifiers, (ent.id, stmt)
+                    assert "P2309" in stmt.qualifiers, (ent.id, stmt)
+                    # and the relation must be instanceof or (instanceof or subclassof), or subclassof
+                    relations = [
+                        x.as_entity_id_safe() for x in stmt.qualifiers["P2309"]
+                    ]
+                    for relation in relations:
+                        assert relation in ["Q21503252", "Q30208840", "Q21514624"], (
+                            ent.id,
+                            stmt,
+                        )
+                except:
+                    continue
+                ranges = [x.as_entity_id_safe() for x in stmt.qualifiers["P2308"]]
 
         inverse_properties = []
         for stmt in ent.props.get("P1696", []):
@@ -80,10 +122,12 @@ class WDProperty(OntologyProperty):
             parents=sorted(parents),
             related_properties=sorted(related_properties),
             equivalent_properties=sorted(equivalent_properties),
-            subjects=sorted(subjects),
+            domains=domains,
+            ranges=ranges,
             inverse_properties=sorted(inverse_properties),
             instanceof=sorted(instanceof),
             ancestors={},
+            constraints=constraints,
         )
 
     def is_object_property(self):
@@ -110,7 +154,8 @@ class WDProperty(OntologyProperty):
             parents=self.parents,
             related_properties=self.related_properties,
             equivalent_properties=self.equivalent_properties,
-            subjects=self.subjects,
+            domains=self.domains,
+            ranges=self.ranges,
             inverse_properties=self.inverse_properties,
             instanceof=self.instanceof,
             ancestors=self.ancestors,
@@ -118,6 +163,33 @@ class WDProperty(OntologyProperty):
 
     def __str__(self):
         return f"{self.label} ({self.id})"
+
+    def to_dict(self):
+        return {
+            "id": self.id,
+            "label": self.label.to_dict(),
+            "description": self.description.to_dict(),
+            "datatype": self.datatype,
+            "aliases": self.aliases.to_dict(),
+            "parents": self.parents,
+            "related_properties": self.related_properties,
+            "equivalent_properties": self.equivalent_properties,
+            "domains": self.domains,
+            "ranges": self.ranges,
+            "inverse_properties": self.inverse_properties,
+            "instanceof": self.instanceof,
+            "ancestors": self.ancestors,
+            "constraints": [s.to_dict() for s in self.constraints],
+        }
+
+    @classmethod
+    def from_dict(cls, obj):
+        obj["label"] = MultiLingualString(**obj["label"])
+        obj["description"] = MultiLingualString(**obj["description"])
+        obj["aliases"] = MultiLingualStringList(**obj["aliases"])
+        obj["ancestors"] = obj["ancestors"]
+        obj["constraints"] = [WDStatement.from_dict(x) for x in obj["constraints"]]
+        return cls(**obj)
 
 
 def normalize_wikidata_datatype(datatype: WDDataType) -> str: